Create SiC_tensile.py
Using base "find peak centres" methodology, reads multiple directories containing a series of .dat files produced from integration of diffraction rings. In this case, the code highlights the peak in each .dat file between channels 1114 and 1164, equivalent to a (1 0 8) diffraction peak for SiC crystals at the relevant energy. The peak is fitted to a two factor gaussian to determine the peak centre, giving a value of the crystallographic d-spacing. This is used to calculate the local tensile strain at any given point, and plotted into strain "maps" for each sub-directory (load stage).
Returns .png images of the smoothed and raw maps, and the data in two .csv dataframes for further analysis