Create Al_tensile.py
Using base "find peak centres" methodology, reads multiple directories containing a series of .dat files produced from integration of diffraction rings. In this case, the code highlights the peak in each .dat file between channels 840 and 880, equivalent to a (2 0 0) diffraction peak for Al crystals at the relevant energy. The peak is fitted to a gaussian distribution to determine the peak centre, giving a value of the crystallographic d-spacing. This is used to calculate the local tensile strain at any given point, and plotted into strain "maps" for each sub-directory (load stage).
Returns .png images of the smoothed and raw maps, and the data in two .csv dataframes for further analysis